refactor: alert list as client component
