+++ src/app/(topics)/(adverse-weather)/adverse-weather/[weather]/page.tsx
@@ -1,7 +1,6 @@
 import { flag } from '@unleash/nextjs'
 import { Metadata } from 'next'
 import Link from 'next/link'
-import { redirect } from 'next/navigation'
 
 import { HealthAlertTypes } from '@/api/models/Alerts'
 import { PageType } from '@/api/requests/cms/getPages'
@@ -10,20 +9,11 @@ import { RelatedLink, RelatedLinks, View } from '@/app/components/ui/ukhsa'
 import HealthAlertsLink from '@/app/components/ui/ukhsa/Links/HealthAlertsLink/HealthAlertsLink'
 import { List } from '@/app/components/ui/ukhsa/List/List'
 import { ListItem } from '@/app/components/ui/ukhsa/List/ListItem'
-import {
-  ListItemStatus,
-  ListItemStatusContent,
-  ListItemStatusIcon,
-  ListItemStatusLink,
-  ListItemStatusTag,
-  ListItemStatusTimestamp,
-} from '@/app/components/ui/ukhsa/List/ListItemStatus'
 import { flags } from '@/app/constants/flags.constants'
-import useWeatherHealthAlertList from '@/app/hooks/queries/useWeatherHealthAlertList'
-import { useTranslation } from '@/app/i18n'
 import { renderCompositeBlock } from '@/app/utils/cms.utils'
 import { extractHealthAlertTypeFromSlug } from '@/app/utils/weather-health-alert.utils'
-import { logger } from '@/lib/logger'
+
+import AlertList from './AlertList'
 
 export async function generateMetadata({ params: { weather } }: { params: { weather: string } }): Promise<Metadata> {
   const { enabled } = await flag(flags.adverseWeather)
@@ -81,15 +71,6 @@ export default async function WeatherHealthAlert({ params: { weather } }: Weathe
 
   const { title, body, related_links: relatedLinks } = await getPageBySlug<PageType.Composite>(weather)
 
-  const healthAlerts = useWeatherHealthAlertList({ type })
-
-  if (healthAlerts.error || !healthAlerts.data) {
-    logger.error(healthAlerts.error)
-    return redirect('/error')
-  }
-
-  const { t } = await useTranslation('adverseWeather')
-
   return (
     <View
       heading={title}
@@ -113,32 +94,7 @@ export default async function WeatherHealthAlert({ params: { weather } }: Weathe
               role="presentation"
             />
 
-            <List>
-              {healthAlerts.data.map(({ status, geography_name: name, refresh_date: lastUpdated }) => (
-                <ListItem key={name} spacing="s">
-                  <ListItemStatus>
-                    <ListItemStatusIcon level={status} type={type} />
-                    <ListItemStatusContent>
-                      <ListItemStatusLink
-                        href={`/adverse-wether/${weather}-health-alerts/${name.toLowerCase().split(' ').join('-')}`}
-                      >
-                        {name}
-                      </ListItemStatusLink>
-
-                      {lastUpdated === null || lastUpdated === '' ? (
-                        <ListItemStatusTimestamp>-</ListItemStatusTimestamp>
-                      ) : (
-                        <ListItemStatusTimestamp>
-                          {t('lastUpdated', { value: new Date(lastUpdated) })}
-                        </ListItemStatusTimestamp>
-                      )}
-                    </ListItemStatusContent>
-
-                    <ListItemStatusTag level={status} type={type} region={name} />
-                  </ListItemStatus>
-                </ListItem>
-              ))}
-            </List>
+            <AlertList type={type} />
           </div>
 
           <h3 className="govuk-heading-m govuk-!-margin-top-8 govuk-!-margin-bottom-1">Further advice and guidance</h3>
